Clean up lexer code.
Since the lexer is a template, it makes sense to have as much code
as possible outside of it.  This is because members of the class
template (including internal classes) that don't use the template
argument require it in their definitions and/or initiallizations
anyway.  This makes the code unnecessarily unreadable.