Improve the implemented parser.
When identifying an operator, instead of "stealing" the next token,
use a flag so that the next token will be placed in the correct
position by the next iteration of the loop.  This makes the code
more compact and takes advantage of the use of semicolons.  This
should also make it easier to generalize the parser for more
complex expresions.